// src/app/shared/models/bgt_kegiatan_cost.model.ts
export class BgtKegiatanCost {
	id?: number;
  kegiatan_id?: number;
  afdeling_id?: number;
  lokasi_id?:number;
  tahun?: string;
  b01?: number;
  b02?: number;
  b03?: number;
  b04?: number;
  b05?: number;
  b06?: number;
  b07?: number;
  b08?: number;
  b09?: number;
  b10?: number;
  b11?: number;
  b12?: number;
}